Update CC1101.cpp
Due to a bug un the CC1101 hardware (see CC1101 erata) a single read of some registers can give incorrect results. The solution is to read these registers twice  (readRegisterWithSyncProblem()) and if the values are the same the result is considered reliable. However, this approach fails for reading the number of bytes in the TX_FIFO during transmission, since this number is continuously updated, also in between the two consequtive reads. Therefore the original statement only reads  the same value twice when the transmisson is completed and the TX_FIFO is empty. But that is too late to write the remaining bytes of the buffer to the TX_FIFO and  the subroutine hangs on a TX_FIFO_BUFFER_UNDERFLOW.
The new statement reads number of bytes in the TX_FIFO as a median of 3 values (readRegisterMedian3()). (High probability that this is correct).  Furthermore, at the end of sendData the state is tested also for a TX_FIFO_BUFFER_UNDERFLOW to prevent a hang.
The function readRegisterMedian3() has been added.