Standby and fullpower fan speed commands added.
The following changes were made:
- standby command added (slowest possible fan speed)
- fullpower command added (fastest possible fan speed)
- message1 is not transmitted to the itho fan anymore since it is not
required
- simplified message1 parsing on receive to support the standard rft
remote and the duco c system remote

The message1 handling can be completely removed from the code. It is not
required for controlling the itho fan. I did not remove the message1
yet, just to leave the options open for future changes and to ease
reverse engineering.